Task 1: New customer creation and total spent after purchases.
A customer's information will be stored in it's own class with their name, email, and a list of their purchase history. Methods will add a purchase amount to the list of purchase history and will then take those amounts from the array and add them up to return the total amount.